+++ TerranEngine/src/Events/KeyboardEvent.h
@@ -7,32 +7,32 @@ namespace TerranEngine
 	class KeyboardEvent : public Event 
 	{
 	public:
-		KeyboardEvent(const int keyCode) 
+		KeyboardEvent(const uint32_t keyCode)
 			: m_KeyCode(keyCode) {}
-		int GetKeyCode() const { return m_KeyCode; }
+		uint32_t GetKeyCode() const { return m_KeyCode; }
 
 		EVENT_CLASS_CATEGORY(EventCategoryKeyboard)
 	private: 
-		int m_KeyCode;
+		uint32_t m_KeyCode;
 	};
 
 	class KeyPressedEvent : public KeyboardEvent 
 	{
 	public:
-		KeyPressedEvent(const int keyCode, const int repeatCount)
+		KeyPressedEvent(const uint32_t keyCode, const uint8_t repeatCount)
 			: KeyboardEvent(keyCode), m_RepeatCount(repeatCount) {}
 
-		int GetRepeatCount() const { return m_RepeatCount; }
+		uint8_t GetRepeatCount() const { return m_RepeatCount; }
 
 		EVENT_CLASS_TYPE(KeyPressed)
 	private:
-		int m_RepeatCount;
+		uint8_t m_RepeatCount;
 	};
 
 	class KeyReleasedEvent : public KeyboardEvent 
 	{
 	public:
-		KeyReleasedEvent(const int keyCode)
+		KeyReleasedEvent(const uint32_t keyCode)
 			: KeyboardEvent(keyCode) {}
 
 		EVENT_CLASS_TYPE(KeyReleased)
